Replace smooth scrolling by jump to anchor for mobile devices
Because on mobile devices smooth scrolling not look great yet. It's
freezing for long jumps to anchor links, and looks not so smooth.

I try to use default handler for mobile devices, but on mobile chrome
and all firefox links do not work.
So i added window.scrollTo for mobile width and on all mobile devices
links works.

iss #2013