Disable scroll when menu appear
Add to body fixed class when menu appear.
Fixed class added overflow hidden and position fixed property
Propery fixed set window to beginning of the page, but after property
disabled we stay on beginnning of the page.
To fixed it we need to save last postion and restore it after scroll
enable

iss #2039